\name{trichoptera}
\alias{trichoptera}
\docType{data}
\title{Trichoptera data set}
\description{
  A transect of 22 "Week" emergence traps, 0.25 m^2 in capture area, was laid along the outflow stream of Lac Cromwell (45o59'13"N, 73o59'47"W) on the territory of Station de Biologie dels Laurentides. The stream was surveyed during a period of 100 days; the traps were set on May and visited daily form May to September 1984. The number of emerging adults belonging to 56 Trichoptera (Insecta) species was noted every day. Insect counts within each 10-day period were pooled. The data set comprises 22 sites, 10 time periods and 56 species.
}
\usage{data(trichoptera)}
\format{
  A data frame with 220 rows and 58 columns. Rows are observations of sites during each time period. There are 10 blocks (time period) of 22 observations (sites). The first two columns indicate the \code{Site} and \code{date}, respectively. The remaining columns correspond to species.
}
\source{
Data set generously donated by Pierre-Paul Harper, Universite de Montreal.
}
\references{
Legendre, P., M. De Caceres and  D. Borcard 2009. Community surveys through space and time to assess environmental changes: testing the space-time interaction in the absence of replication. Ecology (in press).

}
\examples{
data(trichoptera)
}
\keyword{datasets}
